Revert "Tentative change to *Ops. Not final ironoxide_macros"
This reverts commit 7934e5baf845d18145a88bf824322a1affe5467c.
