--- ironoxide_macros/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ironoxide_macros"
-version = "0.1.0"
-edition = '2018'
-
-[dependencies]
-syn = "1.0"
-proc-quote = "0.3.0"
-proc-macro2 = "1.0"
-
-[lib]
-proc-macro = true
